Add ZoneData loading, more World->Zone work

+++ DFEngine/Network/Protocols/Character/PROTO_NC_CHAR_LOGINFAIL_ACK.cs
@@ -2,7 +2,7 @@
 {
 	public class PROTO_NC_CHAR_LOGINFAIL_ACK : NetworkMessage
 	{
-		public PROTO_NC_CHAR_LOGINFAIL_ACK(int error) : base(NetworkCommand.NC_CHAR_LOGINFAIL_ACK)
+		public PROTO_NC_CHAR_LOGINFAIL_ACK(ushort error) : base(NetworkCommand.NC_CHAR_LOGINFAIL_ACK)
 		{
 			Write(error);
 		}
